Add files via upload
the goal :
1- Analysis of customer satisfaction in price versus space.
2- Price trends over time.
3- Demand analysis to determine the most popular type of real estate.
4 - The country to which the customer is linked.
5- Segmenting customers according to purpose.
6 - Analyzing customer satisfaction with the sources through which the deal was made, via the website, the agency, or the client.

Result:
Satisfaction with the price versus the area. The scatter plot shows a positive result in the price of the property according to the area and dividing customers according to the purpose. It leads to the fact that the purpose of the property is to buy a residential house and is estimated at 69% compared to the investment. The most popular country is the United States and an analysis of customer satisfaction from the sources in which the transaction was completed. The site obtained The electronic product has the highest satisfaction rate of 5 out of 5. And the price trends over the years, we note that it witnessed a significant increase in 2007 and a decrease in prices from 2008 to 2010.

Recommendations:
A temporary increase in real estate prices in order to compare the company’s revenues before the real estate price is reduced or increased, and on the basis of this we decide whether prices will rise or not. Providing a questionnaire to customers to measure customer satisfaction with the company and providing recommendations from customers